// src/types/product.type.ts
export interface Product {
	_id: string;
	images: string[];
	price: number;
	rating: number;
	price_before_discount: number;
	quantity: number;
	sold: number;
	view: number;
	name: string;
	description: string;
	category: {
		_id: string;
		name: string;
	};
	image: string;
	createdAt: string;
	updatedAt: string;
}

export interface ProductList {
	products: Product[];
	pagination: {
		page: number;
		limit: number;
		page_size: number;
	};
}

export interface ProductListConfig {
	page?: string;
	limit?: string;
	order?: "asc" | "desc";
	sort_by?: "createdAt" | "view" | "sold" | "price";
	exclude?: string;
	rating_filter?: string;
	price_max?: string;
	price_min?: string;
	name?: string;
	category?: string;
}